refactor(core): split frames pipeline into download/ffmpeg/vision/orchestrator adapters (Phase 3 of #87)
